document_rag_back/requirements.txt
# --- web/api/db/core ---
fastapi
pydantic-settings
sqlalchemy
aiosqlite
python-multipart
uvicorn

# --- vector/RAG ---
chromadb
sentence-transformers
faiss-cpu
langchain
langchain_community

# --- OCR engines & imaging ---
pillow
PyMuPDF
pytesseract

# EasyOCR stack (requires torch + torchvision)
torch==2.4.1
torchvision==0.19.1
torchaudio==2.4.1
easyocr

# PaddleOCR stack (3.x pairs with PaddlePaddle 3.x)
paddlepaddle>=3.0,<3.3
paddleocr>=3.0,<3.1
